arabic_proofreader_backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
PyMuPDF==1.26.3
python-docx==1.1.2
requests==2.32.3
gunicorn==23.0.0

